Add first iteration of mk2 - untested
