+++ index.md
@@ -225,8 +225,8 @@
 
 ## Text
     ViewerNodeTextMK2
-    SvTextInNode
-    SvTextOutNode
+    SvTextInNodeMK2
+    SvTextOutNodeMK2
     NoteNode
     SvDataShapeNode
     GTextNode
